consolidate logic, fix pagination route params
