+++ ui/lib/ldap/addon/routes/roles/subdirectory.ts
@@ -3,21 +3,19 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-import Route from '@ember/routing/route';
+import LdapRolesRoute from '../roles';
 import { hash } from 'rsvp';
 import { ldapBreadcrumbs } from 'ldap/utils/ldap-breadcrumbs';
-import { service } from '@ember/service';
 
 import type { Breadcrumb } from 'vault/vault/app-types';
 import type Controller from '@ember/controller';
 import type LdapRoleModel from 'vault/models/ldap/role';
-import type PaginationService from 'vault/services/pagination';
 import type SecretEngineModel from 'vault/models/secret-engine';
-import type SecretMountPath from 'vault/services/secret-mount-path';
 import type Transition from '@ember/routing/transition';
 
 interface LdapRolesSubdirectoryRouteModel {
   backendModel: SecretEngineModel;
+  parentRole: { path_to_role: string; type: string };
   roles: Array<LdapRoleModel>;
 }
 
@@ -32,10 +30,7 @@ interface LdapRolesSubdirectoryParams {
   type: string;
 }
 
-export default class LdapRolesSubdirectoryRoute extends Route {
-  @service declare readonly pagination: PaginationService;
-  @service declare readonly secretMountPath: SecretMountPath;
-
+export default class LdapRolesSubdirectoryRoute extends LdapRolesRoute {
   queryParams = {
     pageFilter: {
       refreshModel: true,
@@ -48,20 +43,11 @@ export default class LdapRolesSubdirectoryRoute extends Route {
   model(params: LdapRolesSubdirectoryParams) {
     const backendModel = this.modelFor('application') as SecretEngineModel;
     const { path_to_role, type } = params;
-    const page = Number(params.page) || 1;
+    const parentRole = { path_to_role, type };
     return hash({
       backendModel,
-      roles: this.pagination.lazyPaginatedQuery(
-        'ldap/role',
-        {
-          backend: backendModel.id,
-          page,
-          pageFilter: params.pageFilter,
-          responsePath: 'data.keys',
-          skipCache: page === 1,
-        },
-        { adapterOptions: { parentRole: { path_to_role, type } } }
-      ),
+      parentRole,
+      roles: this.lazyQuery({ parentRole }, backendModel.id, params),
     });
   }
 
@@ -71,13 +57,12 @@ export default class LdapRolesSubdirectoryRoute extends Route {
     transition: Transition
   ) {
     super.setupController(controller, resolvedModel, transition);
-    const { backendModel } = resolvedModel;
-    const { path_to_role, type } = this.paramsFor('roles.subdirectory');
+    const { backendModel, parentRole } = resolvedModel;
     const crumbs = [
       { label: 'Secrets', route: 'secrets', linkExternal: true },
       { label: backendModel.id, route: 'overview' },
       { label: 'Roles', route: 'roles' },
-      ...ldapBreadcrumbs(path_to_role, type, backendModel.id),
+      ...ldapBreadcrumbs(parentRole.path_to_role, parentRole.type, backendModel.id),
     ];
 
     // must call 'set' so breadcrumbs update as we navigate through directories
