+++ ui/lib/ldap/addon/templates/roles/index.hbs
@@ -9,4 +9,5 @@
   @backendModel={{this.model.backendModel}}
   @breadcrumbs={{this.breadcrumbs}}
   @pageFilter={{this.pageFilter}}
+  @currentRouteParams={{array this.model.backendModel.id}}
 />
