+++ ui/lib/ldap/addon/templates/roles/subdirectory.hbs
@@ -3,10 +3,13 @@
   SPDX-License-Identifier: BUSL-1.1
 ~}}
 
-<Page::Roles
-  @roles={{this.model.roles}}
-  @promptConfig={{false}}
-  @backendModel={{this.model.backendModel}}
-  @breadcrumbs={{this.breadcrumbs}}
-  @pageFilter={{this.pageFilter}}
-/>
+{{#let this.model.parentRole.type this.model.parentRole.path_to_role as |roleType pathToRole|}}
+  <Page::Roles
+    @roles={{this.model.roles}}
+    @promptConfig={{false}}
+    @backendModel={{this.model.backendModel}}
+    @breadcrumbs={{this.breadcrumbs}}
+    @pageFilter={{this.pageFilter}}
+    @currentRouteParams={{array this.model.backendModel.id roleType pathToRole}}
+  />
+{{/let}}
